refactor: split MySQL port implementations
