+++ src/infra/adapters/mysql/mod.rs
@@ -1,6 +1,8 @@
 mod adapter;
 mod cli;
+mod connection;
 mod dsn;
+mod executor;
 mod metadata;
 mod option_file;
 mod sql;
@@ -8,7 +10,7 @@ mod sql;
 pub use adapter::MySqlAdapter;
 
 #[cfg(all(unix, feature = "test-support"))]
-pub use adapter::test_support::run_mysql_cli_script_for_test;
+pub use executor::test_support::run_mysql_cli_script_for_test;
 
 #[cfg(feature = "test-support")]
-pub use adapter::test_support::export_mysql_csv_to_path_for_test;
+pub use executor::test_support::export_mysql_csv_to_path_for_test;
